The RDF config file are renamed and restructure. As well the UI library and the CSS files. The image upload custom entry form works. It allows the adding of existing images. The only problem that the Document class group page does not work for some reason.
