+++ productMods/config/listViewConfig-depictedIn.xml
@@ -8,18 +8,20 @@
     	PREFIX vitro-public: &lt;http://vitro.mannlib.cornell.edu/ns/vitro/public#&gt;
     	PREFIX rdfs: &lt;http://www.w3.org/2000/01/rdf-schema#&gt;
     	
-		SELECT DISTINCT ?image ?file ?downloadLoc ?downloadUrl ?filename ?fileDescription ?mimeType
+		SELECT DISTINCT ?image ?label ?file ?downloadLoc ?downloadUrl ?filename ?fileDescription ?mimeType
 		WHERE
 		{
 		     ?subject ?property ?image .
-		   	 OPTIONAL {
+			 OPTIONAL {
+			 	?image rdfs:label ?label .
+			 }
+			 OPTIONAL {		   	 
 		     	?image &lt;http://vivo.mydomain.edu/individual/hasFile&gt; ?file .
 		     	?file vitro-public:filename ?filename .
 		     	?file vitro-public:mimeType ?mimeType .
 		     	?file vitro-public:downloadLocation ?downloadLoc .
 		     	?downloadLoc vitro-public:directDownloadUrl ?downloadUrl . 
-		     }
-		     		  
+		     }		  
 		     OPTIONAL { ?file &lt;http://vivo.mydomain.edu/individual/fileDescription&gt; ?fileDescription  }		     
 		}
