--- productMods/css/customEntryForms/instanceOffer.css
@@ -1,32 +1,0 @@
-
-
-.rowContainer {
-	margin : 4px;
-}
-
-.column {
-	width : 140px;
-	overflow : auto;
-	display : inline-block;
-	margin : 5px;
-}
-
-.columnImg {
-	display : inline-block;
-	width : 60px;
-	height : 60px;	
-}
-
-.addInstanceButton {
-	display : inline-block;
-}
-
-.inline {
-	display : inline-block;
-}
-
-.noResult {
-	    font-style: italic;
-	margin : 10px 20px;
-	
-}
