+++ productMods/css/font.css
@@ -1,0 +1,7 @@
+.italic {
+	font-style: italic;
+}
+
+bold {
+    font-weight: bold;
+}
