+++ productMods/css/moduleLoader/View.css
@@ -1,0 +1,40 @@
+.moduleTitle {
+	font-style : italic;	
+	margin : 10px 2px;
+	display : inline-block;
+	font-size : 18px;
+ 	border-bottom : 1px solid;
+	padding : 3px;
+}
+
+.newLine {
+	display : block;
+}
+
+.fullScreenContainer {
+	position: absolute;
+	top: 0;
+	bottom: 0;
+	left: 0;
+	right: 0;
+	overflow: hidden;
+	z-index: 1;
+	background-color: rgba(0, 0, 0, 0.4);
+}
+
+.innerContainer {
+	max-height : 500px;
+	overflow : auto;
+}
+
+.imgContainer {
+	vertical-align : middle;
+	display : inline-block;
+	margin : 3px;
+	width : 16px;
+}
+
+.classLabel {
+	line-height : 20px;
+	display : inline-block;
+}
