+++ productMods/css/table.css
@@ -1,0 +1,64 @@
+.columnTitle {
+	display: inline-block;
+	width: 170px;
+	border-bottom: 1px solid black;
+	margin: 5px;
+	text-align: center;
+}
+
+.rowContainer {
+	display: table;
+	margin: 5px;
+	height: 100px;
+}
+
+.rowInnerContainer {
+	display : flex;
+	align-items : center;
+	height: 100px;
+}
+
+.column {
+	width: 140px;
+	overflow: auto;
+	display: inline-block;
+	height: 100%; 
+	margin : 5px;
+	text-align: center;
+	margin: 5px;
+}
+
+.textColumn {
+	display: flex;
+	height: 100%; align-items : center;
+	overflow: auto;
+	align-items: center;
+}
+
+.noResult {
+	font-style: italic;
+	margin: 10px 20px;
+}
+
+.verticalAlignHelper {
+	display: inline-block;
+	height: 100%;
+	vertical-align: middle;
+}
+
+.buttonCell {
+	display: table-cell;
+	vertical-align: middle;
+}
+
+/****************************************************
+*** Adding gif to the table  
+*****************************************************/
+.dataWaitGif {
+	margin: 5x 40px;
+}
+
+.msgText {
+	max-width : 300px;
+	margin : 20px;
+}
