+++ productMods/js/lib/TableUI.js
@@ -1,0 +1,13 @@
+
+
+var TableUI = {
+
+	getWaitGif : function(){
+		return UI.getGeneralImgSize("loading", "middle").addClass("dataWaitGif")
+	},
+
+	getHelperSpan : function(type, size){
+		
+		return html.span("verticalAlignHelper")
+	}
+}
