+++ rdf/tbox/firsttime/RDFBonesPropertyGroupMembers.n3
@@ -1,8 +1,10 @@
+@prefix vitro:   <http://vitro.mannlib.cornell.edu/ns/vitro/0.7#> .
+
 <http://w3id.org/rdfbones/core#isDepiceted>
 	vitro:inPropertyGroupAnnot
-    	<http://vivoweb.org/ontology#vitroPropertyGroupFile> ;
+    	<http://vivoweb.org/ontology#vitroPropertyGroupFile> .
     						
-			
+
 <http://vivo.mydomain.edu/individual/hasFile>
     	vitro:inPropertyGroupAnnot
     		<http://vivoweb.org/ontology#vitroPropertyGroupFile>.
