+++ rdf/tbox/firsttime/documentDataGetter.n3
@@ -1,5 +1,6 @@
 @prefix display: <http://vitro.mannlib.cornell.edu/ontologies/display/1.1#> .
 @prefix bibo:    <http://purl.org/ontology/bibo/> .
+@prefix vivoweb: <http://vivoweb.org/ontology#> .
 
 
 bibo:Document display:hasDataGetter display:documentsDataGetter.
