+++ rdf/tbox/firsttime/documentVClassGroupClasses.n3
@@ -1,0 +1,15 @@
+
+<http://purl.org/ontology/bibo/Document>
+	vitro:inClassGroup <http://vivoweb.org/ontology#vitroClassGroupdocuments> .
+
+<http://purl.org/ontology/bibo/Image>
+   vitro:inClassGroup <http://vivoweb.org/ontology#vitroClassGroupdocuments> .
+   
+<http://purl.org/ontology/bibo/AudioVisualDocument>
+    vitro:inClassGroup <http://vivoweb.org/ontology#vitroClassGroupdocuments> .
+ 
+<http://purl.org/ontology/bibo/Note>
+ 	   vitro:inClassGroup <http://vivoweb.org/ontology#vitroClassGroupdocuments> .
+ 
+<http://purl.org/ontology/bibo/Slideshow>
+ 	   vitro:inClassGroup <http://vivoweb.org/ontology#vitroClassGroupdocuments> .
